requirements_bridge.txt
pyzmq>=25.0
fastapi>=0.100
uvicorn[standard]>=0.20
